+++ src/templates/semantic/getAlert.js
@@ -1,0 +1,12 @@
+export default function getAlert(type, children) {
+  var className = {
+    ui: true,
+    message: true
+  };
+  className[type] = true;
+  return {
+    tag: 'div',
+    attrs: { className: className },
+    children: children
+  };
+}