+++ src/templates/semantic/getError.js
@@ -1,0 +1,10 @@
+export default function getError(locals) {
+  if (!locals.hasError || !locals.error) { return; }
+  return {
+    tag: 'div',
+    attrs: {
+      className: 'ui pointing label visible red'
+    },
+    children: locals.error
+  };
+}